Relocate HandshakeReceipt to p2p.receipt
